+++ CHANGES.rst
@@ -10,6 +10,7 @@ Upcoming Changes
 
 - Add :py:func:`get_member() <Guild.get_member>` function to :py:class:`Guild`.
 - Add :py:attr:`rank <Member.rank>` property to :py:class:`Member`.
+- Add :py:attr:`best_class <Player.best_class>` property to :py:class:`Player`.
 - :py:func:`OnlinePlayers.get_player_server` and :py:func:`OnlinePlayers.is_player_online` now accept :py:class:`Member` as argument.
 - Support Python 3.10